pyproject.toml
[project]
name = "lwagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "openai>=2.21.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.1",
    "sqlalchemy>=2.0.46",
]
